requirements.txt
autopep8==1.4.3
Django==2.1.3
django-cors-headers==2.4.0
jsonfield==2.0.2
pycodestyle==2.5.0
pytz==2018.7
djangorestframework
python-dateutil
pillow